Move API endpoints from Core to Api project
